拆分设置App
