+++ main/CMakeLists.txt
@@ -4,9 +4,13 @@ idf_component_register(SRCS
                             "apps/pwm_servo/app_pwm_servo.c" 
                             "hal/hal_pwm.c" 
                             "utils/menu_utils.c" 
+                            "utils/memory_utils.c"
                             "managers/gesture_handler.c" 
                             "apps/ark/app_ark.c"
-                            "apps/settings/app_settings.c" 
+                            "apps/settings/app_settings.c"
+                            "apps/settings/settings_about.c"
+                            "apps/settings/settings_display.c"
+                            "apps/settings/settings_sound.c"
                             "managers/app_manager.c" 
                             "apps/launcher/app_launcher.c" 
                             "overlay_drawer.c" 
@@ -24,5 +28,4 @@ idf_component_register(SRCS
                             "assets/image1.c"
                             "assets/scenery1.c"
                             "assets/simhei_32.c"
-                            "utils/memory_utils.c"
                        INCLUDE_DIRS ".")
